==> notebooks/0_evaluate_distance_range.ipynb <==
import math

def side(a,b,alpha):
    return math.sqrt(a**2+b**2-2*a*b*math.cos(math.radians(alpha)))

# O--H
#     N
side(1.01,1.5,120)

# O---H-N
side(1.01,2.5,180)

# O--H-O
side(1.01,2.04,105)

side(1.01,2.04,180)
