==> notebooks/1_generate_pml_scripts.ipynb <==
AT14 = ['4YAY','4ZUD']
AT16 = ['6DO1','6OS0','6OS1','6OS2']
arr = ['6TKO','6U1N','6UP7','6PWC']

filtered_without_arr_path = '../pdb/0_filtered/without_arr'
filtered_arr_path = '../pdb/0_filtered/arr'
filtered_AT1_path = '../pdb/0_filtered/AT1'

BW_without_arr_path = '../pdb/1_BW/without_arr'
BW_arr_path = '../pdb/1_BW/arr'
BW_AT1_path = '../pdb/1_BW/AT1'

aligned_AT1_path = '../pdb/2_aligned/AT1'
aligned_arr_path = '../pdb/2_aligned/arr'
aligned_cutted_arr_path = '../pdb/2_aligned/cutted_arr'

# ! mkdir -p {BW_without_arr_path}
# ! mkdir -p {BW_arr_path}
# ! mkdir -p {filtered_without_arr_path}
# ! mkdir -p {filtered_arr_path}
# ! mkdir -p {filtered_AT1_path}
# ! mkdir -p {aligned_AT1_path}
# ! mkdir -p {aligned_arr_path}
# ! mkdir -p {aligned_cutted_arr_path}

# сшивка пронумерованной BW arr структуры(без arr) с arr

for i in arr:
    with open(f'../scripts/1_BW_arr/{i}.pml', 'w') as f:
        f.write(f'''
reinitialize
load ../{BW_without_arr_path}/{i}_GPCRDB.pdb, BW
load ../{filtered_arr_path}/{i}.pdb, arr
align BW, arr
remove arr and chain R

save ../{BW_arr_path}/{i}.pdb

select round, byres arr around 20 and BW
select orig, byres round around 20 and arr
save ../{aligned_arr_path}/{i}.pdb, (round or orig)
    ''')

# сшивка пронумерованной BW AT1 структуры(без arr) с arr
# изначально расстояние было 4
for AT1 in AT14+AT16:
    for arr_i in arr:
        script = f'''
    reinitialize
    load ../{BW_arr_path}/{arr_i}.pdb, arr
    load ../{BW_AT1_path}/{AT1}_GPCRDB.pdb, AT1
    align AT1, arr
    remove arr and chain R
    alter chain T, chain="R"

    save ../{aligned_AT1_path}/{AT1}_{arr_i}.pdb

    select round, byres arr around 20 and AT1
    select orig, byres round around 20 and arr
    save ../{aligned_cutted_arr_path}/{AT1}_{arr_i}.pdb, (round or orig)
    '''
        with open(f'../scripts/2_align/{AT1}_{arr_i}.pml', 'w') as f:
            f.write(script)
