==> notebooks/2_concat_model_res_dict.ipynb <==
import pandas as pd
from biopandas.pdb import PandasPdb

AT14 = ['4YAY','4ZUD']
AT16 = ['6DO1','6OS0','6OS1','6OS2']
arr = ['6TKO','6U1N','6UP7','6PWC']

total = pd.DataFrame()
for pdb_code in AT14+AT16:
    ppdb=PandasPdb().read_pdb(f'/home/kestr/AT1/pdb/1_BW/AT1/{pdb_code}_GPCRDB.pdb')
    atoms = ppdb.df['ATOM']
    part = atoms.groupby(by = ['chain_id','residue_number']).first().reset_index()[['chain_id','residue_number','residue_name','b_factor']]
    part['pdb_code'] = pdb_code
    total = pd.concat([total,part])
# для общности переименуем b_factor, для тех, у кого нет BW на номер остатка
total.loc[(total.b_factor>46) ,'b_factor']= total[total.b_factor>46]['residue_number']
# проверка того, что по BW отличаются у разных моделей только эти остатки и остатки на краю спиралей
intersect2 = total[total.chain_id=='T'].groupby(['b_factor','residue_name']).count().reset_index()
intersect2[intersect2.pdb_code<4].sort_values(by = 'pdb_code', ascending=False)

for pdb_code in arr:
    ppdb=PandasPdb().read_pdb(f'/home/kestr/AT1/pdb/1_BW/arr/{pdb_code}.pdb')
    atoms = ppdb.df['ATOM']
    part = atoms.groupby(by = ['chain_id','residue_number']).first().reset_index()[['chain_id','residue_number','residue_name','b_factor']]
    part['pdb_code'] = pdb_code
    total = pd.concat([total,part])
# проверка того, что у аррестинов номера отличаются только у крайних остатков
intersect = total[total.chain_id=='A'].groupby(by='residue_number').count()
intersect[intersect.pdb_code<4].sort_values(by = 'pdb_code', ascending=False)

total['chain_id']=total['chain_id'].replace({'T':'R'})

total.to_csv('/home/kestr/AT1/data/model_res_dict.csv', index=False)
