==> notebooks/2_generate_hydrogen_contacts.ipynb <==
import Bio.PDB
import numpy
import pandas as pd

AT14 = ['4YAY','4ZUD']
AT16 = ['6DO1','6OS0','6OS1','6OS2']
arr = ['6TKO','6U1N','6UP7','6PWC']

# min_dist = 2.18
# max_dist = 3.51
min_dist = 0
max_dist = 2
atoms = ['N','O','S','SE']

def calc_atom_dist(atom1, atom2) :
    diff_vector  = atom1.coord - atom2.coord
    return numpy.sqrt(numpy.sum(diff_vector * diff_vector))

def calc_dist_matrix(chain1, chain2) :
    num = 0
    out = pd.DataFrame()
    for res1 in chain1 :
        for atom1 in res1:
            if atom1.element in atoms:
                for res2 in chain2 :
                    for atom2 in res2:
                        if atom2.element in atoms:
                            dist = calc_atom_dist(atom1, atom2)
                            if((dist<max_dist) & (dist>min_dist)):
                                part = pd.DataFrame([
                                    [
                                        chain1.id, num, res1.id[1], atom1.name, dist
                                    ],
                                    [
                                        chain2.id, num, res2.id[1], atom2.name, dist
                                    ]
                                    ],
                                    columns=[
                                        'chain','Cluster_ID', 'res', 'atom', 'dist'
                                    ]
                                )
                                out=pd.concat([out,part])
                                num += 1
    return out

def get_contacts(pdb_code):
    pdb_filename = f"/home/kestr/AT1/pdb/2_aligned/arr/{pdb_code}.pdb"
    structure = Bio.PDB.PDBParser().get_structure(pdb_code, pdb_filename)
    model = structure[0]
    out = calc_dist_matrix(model["A"], model["R"])
    out['gpcr']=pdb_code
    out['arr']=pdb_code
    return out

def get_contacts2(pdb_code1,pdb_code2):
    pdb_filename = f"/home/kestr/AT1/pdb/2_aligned/cutted_arr/{pdb_code1}_{pdb_code2}.pdb"
    structure = Bio.PDB.PDBParser().get_structure(pdb_code, pdb_filename)
    model = structure[0]
    out = calc_dist_matrix(model["A"], model["R"])
    out['gpcr']=pdb_code1
    out['arr']=pdb_code2
    return out

total = pd.DataFrame()
for pdb_code in arr:
    total = pd.concat([total,get_contacts(pdb_code)])

for AT1 in AT14+AT16:
    for arr_i in arr:
        total = pd.concat([total,get_contacts2(AT1,arr_i)])

total[['Cluster_ID','res']]=total[['Cluster_ID','res']].astype('int')

total

total.to_csv('/home/kestr/AT1/data/hydrogen_contacts.csv', index=False)

# total.to_csv('/home/kestr/AT1/data/charge_repulsion.csv', index=False)
