==> notebooks/3_create_table_contacts.ipynb <==
import pandas as pd
import numpy as np

# выводит таблицу в таком формате: столбцы - имя gpcr, строки - BW этих gpsr(или номер остатка, для тех, у кого нет BW), а значения - минимальный номер остатка аррестина.
def table(repulsion,total):
    repulsion_R = (repulsion[repulsion.chain == 'R']
                   .merge(total[total.chain_id == 'R'], left_on=['gpcr','res'],right_on=['pdb_code','residue_number'])
                   )
    repulsion_A = (repulsion[repulsion.chain == 'A']
                   .merge(total[total.chain_id == 'A'], left_on=['arr','res'],right_on=['pdb_code','residue_number']))
    merged_repulsion = repulsion_R.merge(repulsion_A,on=['gpcr','arr','Cluster_ID'])
    total_a=total[total['chain_id']=='A'][['residue_number','residue_name']].drop_duplicates().rename(columns={'residue_number':'res_y','residue_name':'arr_residue_name'})
    b_factor_x = merged_repulsion.pivot_table(values='b_factor_x',index='res_y',columns='gpcr',aggfunc=np.min).astype('Float32').astype('str').replace('<NA>', '').reset_index().merge(total_a,on='res_y').rename(columns={'res_y':'arr_residue_number'})
    residue_name_x = merged_repulsion.pivot_table(values='residue_name_x',index='res_y',columns='gpcr',aggfunc=np.min).astype('str').replace('nan', '').reset_index().merge(total_a,on='res_y').add_suffix('_name').rename(columns={'res_y_name':'arr_residue_number'}).drop(columns=['arr_residue_name_name'])
    dist_x = merged_repulsion.pivot_table(values='dist_x',index='res_y',columns='gpcr',aggfunc=np.min).astype('Float32').round(2).astype('str').replace('<NA>', '').reset_index().merge(total_a,on='res_y').add_suffix('_dist').rename(columns={'res_y_dist':'arr_residue_number'}).drop(columns=['arr_residue_name_dist'])
    merged = b_factor_x .merge(residue_name_x, on = 'arr_residue_number') .merge(dist_x, on = 'arr_residue_number')
    return  merged.reindex(sorted(merged.columns), axis=1).drop_duplicates(subset='arr_residue_number').reset_index(drop=True)

def table_simple(repulsion,total):
    repulsion_R = (repulsion[repulsion.chain == 'R']
                   .merge(total[total.chain_id == 'R'], left_on=['gpcr','res'],right_on=['pdb_code','residue_number'])
                   )
    repulsion_A = (repulsion[repulsion.chain == 'A']
                   .merge(total[total.chain_id == 'A'], left_on=['arr','res'],right_on=['pdb_code','residue_number']))
    merged_repulsion = repulsion_R.merge(repulsion_A,on=['gpcr','arr','Cluster_ID'])
    total_a=total[total['chain_id']=='A'][['residue_number','residue_name']].drop_duplicates().rename(columns={'residue_number':'res_y','residue_name':'arr_residue_name'})
    b_factor_x = merged_repulsion.pivot_table(values='b_factor_x',index='res_y',columns='gpcr',aggfunc=np.min).astype('Float32').astype('str').replace('<NA>', '').reset_index().merge(total_a,on='res_y').rename(columns={'res_y':'arr_residue_number'})
    residue_name_x = merged_repulsion.pivot_table(values='residue_name_x',index='res_y',columns='gpcr',aggfunc=np.min).astype('str').replace('nan', '').reset_index().merge(total_a,on='res_y').add_suffix('_name').rename(columns={'res_y_name':'arr_residue_number'}).drop(columns=['arr_residue_name_name'])
    merged = b_factor_x .merge(residue_name_x, on = 'arr_residue_number')
    return  merged.reindex(sorted(merged.columns), axis=1).drop_duplicates(subset='arr_residue_number').reset_index(drop=True)

AT14 = ['4YAY','4ZUD']
AT16 = ['6DO1','6OS0','6OS1','6OS2']
AT1 = AT14 + AT16
arr = ['6TKO','6U1N','6UP7','6PWC']

hyd_aci = pd.read_csv('/home/kestr/AT1/data/hydrogen_contacts.csv')
repulsion = pd.read_csv('/home/kestr/AT1/data/charge_repulsion.csv')
other_contacts = pd.read_csv('/home/kestr/AT1/data/contacts_ion_and_hydrophobic.csv')
total = pd.read_csv('/home/kestr/AT1/data/model_res_dict.csv')
hydrophob = other_contacts[other_contacts.type=='PH'].drop(columns=['type'])
ionic = other_contacts[other_contacts.type=='IO'].drop(columns=['type'])

total.query('b_factor == 3.54')

# водородные связи
table(hyd_aci,total).to_csv('/home/kestr/AT1/data/table_contacts/HY.csv',index=False)
table_simple(hyd_aci,total)

# стерическое отталкивание
table(repulsion,total).to_csv('/home/kestr/AT1/data/table_contacts/RE.csv',index=False)
table(repulsion,total)

# гидрофобное притяжение
table_simple(hydrophob,total).to_csv('/home/kestr/AT1/data/table_contacts/PH.csv',index=False)
table_simple(hydrophob,total)

# ионное притяжение
table_simple(ionic,total).to_csv('/home/kestr/AT1/data/table_contacts/IO.csv',index=False)
table_simple(ionic,total)
